LS: Move IDE-style feature modules to `ide` module
commit-id:6356d6ca
