+++ crates/cairo-lang-language-server/src/ide/completions.rs
