+++ crates/cairo-lang-language-server/src/ide/mod.rs
@@ -1,0 +1,2 @@
+pub mod completions;
+pub mod semantic_highlighting;
