+++ crates/cairo-lang-language-server/src/ide/semantic_highlighting/encoder.rs
